+++ src/apps/settings/modules/shared/store/domain.ts
@@ -1,5 +1,5 @@
 import { ConnectedMonitor, IUIColors } from '@seelen-ui/lib';
-import { AppConfig, Placeholder, Plugin, Profile, Settings, Theme, Widget, WindowManagerLayout } from '@seelen-ui/lib/types';
+import { AppConfig, IconPack, Placeholder, Plugin, Profile, Settings, Theme, Widget, WindowManagerLayout } from '@seelen-ui/lib/types';
 
 import { Route } from '../../../components/navigation/routes';
 
@@ -10,10 +10,10 @@ export interface RootState extends Settings {
   toBeRestarted: boolean;
   appsConfigurations: AppConfig[];
   availableThemes: Theme[];
+  availableIconPacks: IconPack[];
   availableLayouts: WindowManagerLayout[];
   availablePlaceholders: Placeholder[];
   autostart: boolean | null;
-  wallpaper: string | null;
   colors: IUIColors;
   plugins: Plugin[];
   widgets: Widget[];