+++ src/apps/settings/modules/shared/store/infra.ts
@@ -2,6 +2,7 @@ import { configureStore } from '@reduxjs/toolkit';
 import {
   AppConfigurationList,
   ConnectedMonitorList,
+  IconPackList,
   PlaceholderList,
   PluginList,
   ProfileList,
@@ -10,6 +11,7 @@ import {
   ThemeList,
   UIColors,
   WidgetList,
+  WindowManagerLayoutList,
 } from '@seelen-ui/lib';
 import { listen as listenGlobal } from '@tauri-apps/api/event';
 import { Modal } from 'antd';
@@ -18,11 +20,11 @@ import { cloneDeep } from 'lodash';
 import { startup } from '../tauri/infra';
 
 import { RootActions, RootReducer } from './app/reducer';
-import { StateToJsonSettings, StaticSettingsToState } from './app/StateBridge';
+import { StateToJsonSettings } from './app/StateBridge';
 
 import { RootState } from './domain';
 
-import { saveUserSettings, UserSettingsLoader } from './storeApi';
+import { saveUserSettings } from './storeApi';
 
 const IsSavingSettings = { current: false };
 
@@ -61,6 +63,10 @@ export async function registerStoreEvents() {
     store.dispatch(RootActions.setAvailablePlaceholders(list.all()));
   });
 
+  WindowManagerLayoutList.onChange((list) => {
+    store.dispatch(RootActions.setAvailableLayouts(list.all()));
+  });
+
   ThemeList.onChange((list) => {
     store.dispatch(RootActions.setAvailableThemes(list.all()));
   });
@@ -86,6 +92,10 @@ export async function registerStoreEvents() {
     store.dispatch(RootActions.setState(newState));
   });
 
+  await IconPackList.onChange((list) => {
+    store.dispatch(RootActions.setAvailableIconPacks(list.all()));
+  });
+
   await PluginList.onChange((list) => {
     store.dispatch(RootActions.setPlugins(list.all()));
   });
@@ -102,21 +112,25 @@ export const LoadSettingsToStore = async (customPath?: string) => {
     store.dispatch(RootActions.setAutostart(value));
   });
 
-  const userSettings = await new UserSettingsLoader()
-    .withLayouts()
-    .withPlaceholders()
-    .withUserApps()
-    .withThemes()
-    .withSystemWallpaper()
-    .load(customPath);
-
+  const settings: Settings = customPath ? await Settings.loadCustom(customPath) : await Settings.getAsync();
   const currentState = store.getState();
-  const newState = StaticSettingsToState(userSettings, currentState);
-  store.dispatch(RootActions.setState(newState));
+  store.dispatch(RootActions.setState({
+    ...currentState,
+    ...settings.inner,
+  }));
+
+  store.dispatch(RootActions.setAppsConfigurations((await AppConfigurationList.getAsync()).all()));
+
+  store.dispatch(RootActions.setAvailableThemes((await ThemeList.getAsync()).all()));
+  store.dispatch(RootActions.setAvailableIconPacks((await IconPackList.getAsync()).all()));
+
+  store.dispatch(RootActions.setAvailablePlaceholders((await PlaceholderList.getAsync()).all()));
+  store.dispatch(RootActions.setAvailableLayouts((await WindowManagerLayoutList.getAsync()).all()));
 
   store.dispatch(RootActions.setPlugins((await PluginList.getAsync()).all()));
   store.dispatch(RootActions.setWidgets((await WidgetList.getAsync()).all()));
   store.dispatch(RootActions.setProfiles((await ProfileList.getAsync()).all()));
+
   setMonitorsOnState(await ConnectedMonitorList.getAsync());
 
   const state = { ...store.getState() };