+++ src/background/exposed.rs
@@ -192,6 +192,7 @@ pub fn register_invoke_handler(app_builder: Builder<Wry>) -> Builder<Wry> {
         // Seelen Settings
         set_auto_start,
         get_auto_start_status,
+        state_get_icon_packs,
         state_get_themes,
         state_get_placeholders,
         state_get_layouts,