+++ changelog.md
@@ -4,6 +4,7 @@
 ### features
 - add kill process option on context menu for dock items.
 - multi-language calendar and date module on toolbar.
+- add icon packs selector on settings > general.
 
 ### refactor
 - move dock state to the background.